requirements.txt
Flask
psycopg2
python-decouple

